Moves pdf, epub, html functions to index.js
